+++ packages/saber/lib/Pages.js
@@ -5,6 +5,7 @@ const { log } = require('saber-log')
 const merge = require('lodash.merge')
 const getPermalink = require('./utils/getPermalink')
 const getPageType = require('./utils/getPageType')
+const { prefixAssets } = require('./utils/assetsAttribute')
 
 module.exports = class Pages extends Map {
   constructor(api) {
@@ -117,6 +118,10 @@ module.exports = class Pages extends Map {
       throw new Error(`Page must have an internal id.`)
     }
 
+    page.attributes.assets = page.attributes.assets
+      ? prefixAssets(page.attributes.assets)
+      : {}
+
     // Ensure this page is not saved
     // So that it will be emitted to disk later in `emitPages` hook
     page.internal.saved = false
@@ -148,12 +153,6 @@ module.exports = class Pages extends Map {
       this.pageProps.get(id),
       this.getPagePublicFields(id)
     )
-  }
-
-  extendPageProp(id, page) {
-    this.pageProps.set(id, Object.assign({}, this.pageProps.get(id), page))
-    // Mark this page as unsaved when the page prop changes
-    this.get(id).internal.saved = false
   }
 
   getPagePublicFields(page) {